requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
jinja2==3.1.4
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
openai==1.58.1
stripe==11.4.0
reportlab==4.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==11.0.0
matplotlib==3.10.8
numpy==2.4.4
qrcode==8.2
